054: Extract mixins for Input/Mouse, Input/Keyboard, Environment, and Structure.
