+++ lib/flutter_processing.dart
@@ -2,4 +2,6 @@ library flutter_processing;
 
 export 'src/_core.dart';
 export 'src/color/_color.dart';
+export 'src/input/mouse_buttons.dart';
+export 'src/output/image_file_formats.dart';
 export 'src/shape/shapes.dart';
