+++ lib/src/input/_keyboard.dart
@@ -1,0 +1,10 @@
+part of '../_core.dart';
+
+mixin SketchInputKeyboard {
+  Set<LogicalKeyboardKey> _pressedKeys = {};
+
+  bool get isKeyPressed => _pressedKeys.isNotEmpty;
+
+  LogicalKeyboardKey? _key;
+  LogicalKeyboardKey? get key => _key;
+}
