+++ lib/src/input/mouse_buttons.dart
@@ -1,0 +1,5 @@
+enum MouseButton {
+  left,
+  center,
+  right,
+}
