+++ lib/src/output/image_file_formats.dart
@@ -1,0 +1,6 @@
+enum ImageFileFormat {
+  png,
+  jpeg,
+  tiff,
+  targa,
+}
