+++ lib/src/structure/_structure.dart
@@ -1,0 +1,17 @@
+part of '../_core.dart';
+
+mixin SketchStructure {
+  bool _isLooping = true;
+  VoidCallback? _loop;
+  VoidCallback? _noLoop;
+
+  void loop() {
+    _isLooping = true;
+    _loop?.call();
+  }
+
+  void noLoop() {
+    _isLooping = false;
+    _noLoop?.call();
+  }
+}
